FIX_COUNT_DEFEATED_ENEMIES
Fixes the glitch in the original game where the subroutine CountDefeatedEnemies checks for the number of ally slots available in battle rather than the number of monster slots. This has no effect in the original game for battles 16 and 20 because the number of ally and monster slots are identical. But it makes battle 21 (Chess) much less aggressive than originally designed.